% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divi_admission_federal_state.R
\name{divi_admissions_federal_state}
\alias{divi_admissions_federal_state}
\title{Import DIVI ICU admission data on federal state level}
\usage{
divi_admissions_federal_state(federal_state, save_dataset, save_plot)
}
\arguments{
\item{federal_state}{"A string vector specifying the federal state you are interested in. Use "all" if you want to obtain data for all 16 federal states.}

\item{save_plot}{"TRUE/FALSE"}

\item{save_data_set}{"TRUE/FAlSE"}
}
\value{
A list: 1 data frame containing the reported ICU admissions from end of July 2021 onwards, 2 admissions plot over time.
}
\description{
Import DIVI ICU admission data on federal state level
}
\examples{
Berlin <- divi_admissions_federal_state("Berlin", FALSE, FALSE)
}
