% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rki_incidence_county.R
\name{rki_incidence_county}
\alias{rki_incidence_county}
\title{Incidence over time on county level}
\usage{
rki_incidence_county(county, save_data_set, save_plot)
}
\arguments{
\item{county}{"county or "all""}

\item{save_data_set}{"TRUE/FALSE"}

\item{save_plot}{"TRUE/FALSE, TRUE only possible if county != all"}
}
\value{
A list: 1) data frame containing the incidences for the chosen county, 2) If county != "all", line plot of this incidence over time over time
}
\description{
Incidence over time on county level
}
\examples{
Leipzig <- rki_incidence_county("Leipzig", FALSE, FALSE)
}
