% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceDataRKIFederalStates.R
\name{rki_incidence_federal_state}
\alias{rki_incidence_federal_state}
\title{Import RKI incidence data on federal state level}
\usage{
rki_incidence_federal_state(federal_state, save_data_set, save_plot)
}
\arguments{
\item{federal_state}{"A string vector specifying the federal state you are interested in. Use "all" if you want to obtain data for all 16 federal states.}

\item{save_data_set}{"TRUE/FAlSE"}

\item{save_plot}{"TRUE/FALSE"}
}
\value{
A list: 1 data frame containing the reported incidence from May 2020 onwards, 2 If federal_state == "all", you may find a facet plot of the incidences over time here, else : empty; 3 Incidence plot over time # nolint # nolint
}
\description{
Import RKI incidence data on federal state level
}
\examples{
Berlin <- rki_incidence_federal_state("Berlin", FALSE, FALSE)
}
