% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rki_vaccination_quantities_and_shares.R
\name{rki_vaccination_quantities_and_shares}
\alias{rki_vaccination_quantities_and_shares}
\title{Vaccination counts and manufacturer shares on federal state level}
\usage{
rki_vaccination_quantities_and_shares(
  federal_state,
  save_data_sets,
  save_plots
)
}
\arguments{
\item{federal_state}{"A string vector specifying the federal state you are interested in."}

\item{save_plots}{"TRUE/FALSE"}

\item{save_data_set}{"TRUE/FAlSE"}
}
\value{
A list: 1) data frame containing the weekly vaccination counts for the selected state, 2) counts containing the shares of the different manufacturers, 3) Plot of counts over time, 4) Plots of shares
}
\description{
Vaccination counts and manufacturer shares on federal state level
}
\examples{
vaccinationCountsAndShares <- rki_vaccination_quantities_and_shares("Bayern", FALSE, FALSE)
}
