% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rki_vocs.R
\name{rki_vocs}
\alias{rki_vocs}
\title{VOC shares on national level}
\usage{
rki_vocs(min_date, max_date, how_many)
}
\arguments{
\item{min_date}{"Date"}

\item{max_date}{"Date"}

\item{how_many}{"# of VOCs"}
}
\value{
A list: 1) data frame containing the shares of the how_many most common VOCs, 2) line plot of these VOCs over time
}
\description{
VOC shares on national level
}
\examples{
rkiVOCs <- rki_vocs("2022-01-01", today(), 10)
}
