% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceDataRKIFederalStates.R
\name{rkiIncidenceFederalState}
\alias{rkiIncidenceFederalState}
\title{Import RKI incidence data on federal state level}
\usage{
rkiIncidenceFederalState(federalState, saveDataSet, savePlot)
}
\arguments{
\item{federalState}{"A string vector specifying the federal state you are interested in. Use "all" if you want to obtain data for all 16 federal states.}

\item{saveDataSet}{"TRUE/FAlSE"}

\item{savePlot}{"TRUE/FALSE"}
}
\value{
A list: 1 data frame containing the reported incidence from May 2020 onwards, 2 If federalState == "all", you may find a facet plot of the incidences over time here, else : empty; 3 Incidence plot over time
}
\description{
Import RKI incidence data on federal state level
}
\examples{
Berlin <- rkiIncidenceFederalState("Berlin", FALSE, FALSE)
}
